+++ source/hooks/chat-handler/conversation/tool-executor.tsx
@@ -1,4 +1,4 @@
-import type {ConversationStateManager} from '@/app/utils/conversationState';
+import type {ConversationStateManager} from '@/app/utils/conversation-state';
 import {ErrorMessage} from '@/components/message-box';
 import type {ToolManager} from '@/tools/tool-manager';
 import type {ToolCall, ToolResult} from '@/types/core';
